+++ src/renderer/editor/toolbar/toolbar-item-dragger.ts
@@ -15,12 +15,15 @@ import { ItemDragger } from '../common/item-dragger';
  * drag operation. The `dropTargets` are the elements which are considered drop targets.
  * If the dragged div is dropped onto one of these elements, the `drop` event is emitted.
  */
-export interface DraggedItemInfo {
+export interface DraggedItemInfo<T> {
   /** This is used to identify the dragged item. */
-  index: number;
+  data: T;
 
   /** The class name which will be added to the container during the drag operation. */
   dragClass: string;
+
+  /** If `true`, there will remain a ghost of the dragged item at the original position. */
+  ghostMode: boolean;
 
   /** The elements which are considered drop targets. */
   dropTargets: HTMLElement[];
@@ -35,10 +38,10 @@ export interface DraggedItemInfo {
  * emitted by the `ItemDragger` class:
  *
  * @fires drop - When an item is successfully dropped onto a drop target, this event is
- *   emitted. The event data is the `index` which was passed to the `addDraggable` method
+ *   emitted. The event data is the `data` which was passed to the `addDraggable` method
  *   and the `dropTarget` which was hovered when the item was dropped.
  */
-export class ToolbarItemDragger extends ItemDragger<DraggedItemInfo> {
+export class ToolbarItemDragger<T> extends ItemDragger<DraggedItemInfo<T>> {
   /**
    * This constructor creates a new ToolbarItemDragger. It will store a reference to the
    * #kando-editor element, so this element must exist when the constructor is called.
@@ -49,61 +52,69 @@ export class ToolbarItemDragger extends ItemDragger<DraggedItemInfo> {
     // Elements will be appended to this container during the drag operation.
     const dragContainer = document.querySelector('#kando-editor') as HTMLElement;
 
-    // During drag'n'drop operations, we need to append the dragged div to the outer
-    // container to be able to drag it outside of any scrollable area.
-    let originalParent: HTMLElement;
+    let ghostDiv: HTMLElement;
 
     this.on('drag-start', (info, div) => {
-      // Add a class to the container. This can be used to highlight the drop targets..
+      // Add a class to the container. This can be used to highlight the drop targets.
       dragContainer.classList.add(info.dragClass);
 
-      // Set fixed width and height for dragged item.
+      // Create a clone of the div. We will drag this clone round. We make it use the same
+      // size as the original div.
+      ghostDiv = div.cloneNode(true) as HTMLElement;
       const rect = div.getBoundingClientRect();
-      div.style.width = `${rect.width}px`;
-      div.style.height = `${rect.height}px`;
+      ghostDiv.style.width = `${rect.width}px`;
+      ghostDiv.style.height = `${rect.height}px`;
 
-      // Remember the original parent so that we can reinsert the div later.
-      originalParent = div.parentNode;
+      // Make the original div invisible.
+      div.style.opacity = info.ghostMode ? '0.2' : '0';
 
       // Append the div to the outer container. This is necessary because the div may be
       // inside a scrollable container with overflow: hidden set and we want to be able to
       // drag it outside.
-      div.classList.add('dragging');
-      dragContainer.appendChild(div);
+      ghostDiv.classList.add('dragging');
+      dragContainer.appendChild(ghostDiv);
     });
 
     // During the drag operation, we need to move the div to the current mouse position.
     this.on('drag-move', (info, div, relative, absolute, offset, grabOffset) => {
-      div.style.transform = `translate(${absolute.x - grabOffset.x}px, ${absolute.y - grabOffset.y}px)`;
+      ghostDiv.style.transform = `translate(${absolute.x - grabOffset.x}px, ${absolute.y - grabOffset.y}px)`;
     });
 
     // If the drag is canceled or ends, we need to clean up.
-    const onDragEnd = (info: DraggedItemInfo, div: HTMLElement) => {
-      div.classList.remove('dragging');
-      div.style.transform = '';
 
-      // Clear the fixed width and height.
-      div.style.width = '';
-      div.style.height = '';
-
-      // Reinsert the div back to its original position.
-      originalParent.appendChild(div);
-
+    const dragEnd = (info: DraggedItemInfo<T>, div: HTMLElement, animate: boolean) => {
       dragContainer.classList.remove(info.dragClass);
+      if (animate) {
+        const rect = div.getBoundingClientRect();
+        ghostDiv.classList.remove('dragging');
+        ghostDiv.style.transform = `translate(${rect.left}px, ${rect.top}px)`;
+
+        setTimeout(() => {
+          ghostDiv.remove();
+          div.style.opacity = '1';
+        }, 200);
+      } else {
+        ghostDiv.remove();
+        div.style.opacity = '1';
+      }
     };
 
-    this.on('drag-cancel', onDragEnd);
+    this.on('drag-cancel', (info, div) => {
+      dragEnd(info, div, true);
+    });
 
     // If the drag ends successfully, we emit the 'drop' event if one of the drop targets
     // is hovered.
     this.on('drag-end', (info, div) => {
-      onDragEnd(info, div);
+      let success = false;
 
       info.dropTargets.forEach((element: HTMLElement) => {
         if (element.matches(':hover')) {
-          this.emit('drop', info.index, element);
+          this.emit('drop', info.data, element);
+          success = true;
         }
       });
+      dragEnd(info, div, !success);
     });
   }
 }